requirements.txt
numpy
tensorflow-macos
tensorflow-metal
scikit-learn
pillow
jupyterlab
